# openapi.yaml
openapi: 3.0.3
info:
  title: CodeOps-Server API
  description: >-
    REST API for CodeOps — an AI-powered software maintenance platform.
    Provides QA job orchestration, agent-based code analysis, compliance tracking,
    dependency scanning, tech debt management, health monitoring, and team collaboration.
  version: 1.0.0
servers:
  - url: http://localhost:8090
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      operationId: register
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Authenticate user and obtain JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshToken
      summary: Refresh an expired access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Logout and blacklist the current JWT token
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Logged out successfully
  /api/v1/auth/change-password:
    post:
      tags: [Authentication]
      operationId: changePassword
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully

  # ──────────────────────────────────────────────
  # MFA (Multi-Factor Authentication)
  # ──────────────────────────────────────────────
  /api/v1/auth/mfa/setup:
    post:
      tags: [Authentication]
      operationId: mfaSetup
      summary: Initiate MFA setup (generates TOTP secret and recovery codes)
      description: >-
        Requires authentication. The user must provide their current password.
        Returns a TOTP secret, a QR code data URI, and 8 one-time recovery codes.
        MFA is NOT enabled until verified via /mfa/verify.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaSetupRequest'
      responses:
        '200':
          description: MFA setup initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaSetupResponse'
        '400':
          description: MFA already enabled or wrong password
  /api/v1/auth/mfa/verify:
    post:
      tags: [Authentication]
      operationId: mfaVerify
      summary: Verify a TOTP code and enable MFA
      description: >-
        Requires authentication. Second step of MFA setup — submits a TOTP code
        from the authenticator app to activate MFA on the account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: MFA enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaStatusResponse'
        '400':
          description: Invalid TOTP code or MFA not set up
  /api/v1/auth/mfa/login:
    post:
      tags: [Authentication]
      operationId: mfaLogin
      summary: Complete MFA-challenged login with TOTP code, email code, or recovery code
      security: []
      description: >-
        Publicly accessible. Submits the MFA challenge token (from login) along
        with a 6-digit TOTP code, 6-digit email code, or 8-digit recovery code.
        On success, returns full access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaLoginRequest'
      responses:
        '200':
          description: MFA verified, tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid challenge token or MFA code
  /api/v1/auth/mfa/disable:
    post:
      tags: [Authentication]
      operationId: mfaDisable
      summary: Disable MFA for the current user
      description: >-
        Requires authentication. Clears the TOTP secret, email codes, and recovery codes.
        Resets the MFA method to NONE. The user must provide their current password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaSetupRequest'
      responses:
        '200':
          description: MFA disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaStatusResponse'
        '400':
          description: MFA not enabled or wrong password
  /api/v1/auth/mfa/recovery-codes:
    post:
      tags: [Authentication]
      operationId: mfaRegenerateRecoveryCodes
      summary: Regenerate MFA recovery codes
      description: >-
        Requires authentication. MFA must be enabled. Replaces all existing
        recovery codes with a new set of 8 codes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaSetupRequest'
      responses:
        '200':
          description: New recovery codes generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaRecoveryResponse'
        '400':
          description: MFA not enabled or wrong password
...
  /api/v1/auth/mfa/setup/email:
    post:
      tags: [Authentication]
      operationId: emailMfaSetup
      summary: Initiate email MFA setup (generates recovery codes and sends verification code)
      description: >-
        Requires authentication. The user must provide their current password.
        Generates 8 one-time recovery codes, sets the MFA method to EMAIL, and sends a
        6-digit verification code to the user's registered email.
        MFA is NOT enabled until verified via /mfa/verify-setup/email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaEmailSetupRequest'
      responses:
        '200':
          description: Email MFA setup initiated, recovery codes returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaRecoveryResponse'
        '400':
          description: MFA already enabled or wrong password
  /api/v1/auth/mfa/verify-setup/email:
    post:
      tags: [Authentication]
      operationId: emailMfaVerifySetup
      summary: Verify email code and enable email MFA
      description: >-
        Requires authentication. Second step of email MFA setup — submits the 6-digit
        code received via email to activate email MFA on the account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: Email MFA enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MfaStatusResponse'
        '400':
          description: Invalid code or email MFA setup not initiated
  /api/v1/auth/mfa/resend:
    post:
      tags: [Authentication]
      operationId: mfaResendCode
      summary: Resend MFA login code to the user's email
      security: []
      description: >-
        Publicly accessible. Sends a new 6-digit MFA code to the user's email during
        the email MFA login flow. Requires a valid MFA challenge token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaResendRequest'
      responses:
        '200':
          description: Code resent
        '400':
          description: Invalid challenge token or email MFA not enabled

  # ──────────────────────────────────────────────
  # Admin
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ──────────────────────────────────────────────
  # Reusable Parameters
  # ──────────────────────────────────────────────
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        displayName:
          type: string
          maxLength: 100
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    # ─── MFA Schemas ──────────────────────────────
    MfaSetupRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: Current account password for re-verification
    MfaVerifyRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: 6-digit TOTP code from authenticator app, or 6-digit email verification code
    MfaLoginRequest:
      type: object
      required: [challengeToken, code]
      properties:
        challengeToken:
          type: string
          description: MFA challenge JWT token received from login
        code:
          type: string
          description: 6-digit TOTP code, 6-digit email code, or 8-digit recovery code
    MfaSetupResponse:
      type: object
      properties:
        secret:
          type: string
          description: TOTP secret key (Base32 encoded)
        qrCodeUri:
          type: string
          description: Data URI of QR code PNG image for scanning
        recoveryCodes:
          type: array
          items:
            type: string
          description: 8 one-time recovery codes (8 digits each)
    MfaStatusResponse:
      type: object
      properties:
        mfaEnabled:
          type: boolean
        mfaMethod:
          type: string
          enum: [NONE, TOTP, EMAIL]
          description: The active MFA method
        recoveryCodesRemaining:
          type: integer
          description: Number of unused recovery codes remaining (null when MFA disabled)
          nullable: true
    MfaEmailSetupRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: Current account password for re-verification
    MfaResendRequest:
      type: object
      required: [challengeToken]
      properties:
        challengeToken:
          type: string
          description: MFA challenge JWT token received from login
    MfaRecoveryResponse:
      type: object
      properties:
        recoveryCodes:
          type: array
          items:
            type: string
          description: 8 newly generated one-time recovery codes
...
    AuthResponse:
      type: object
      description: >-
        Authentication response. For normal logins, contains token, refreshToken, and user.
        When MFA is required, contains mfaRequired=true and mfaChallengeToken instead.
        For email MFA challenges, also includes maskedEmail.
      properties:
        token:
          type: string
          description: JWT access token (null when MFA challenge is required)
        refreshToken:
          type: string
          description: JWT refresh token (null when MFA challenge is required)
        user:
          $ref: '#/components/schemas/UserResponse'
        mfaRequired:
          type: boolean
          description: True when the user has MFA enabled and must complete verification
        mfaChallengeToken:
          type: string
          description: Short-lived JWT challenge token for MFA login (5 min expiry)
        maskedEmail:
          type: string
          description: Masked email address for email MFA challenges (e.g., "a***@example.com")
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
        mfaEnabled:
          type: boolean
          description: Whether multi-factor authentication is enabled on this account
        mfaMethod:
          type: string
          enum: [NONE, TOTP, EMAIL]
          description: The active MFA method for this account
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...